+++ temba/msgs/tests.py
@@ -31,7 +31,7 @@ from temba.tests import AnonymousOrg, CRUDLTestMixin, TembaTest, mock_uuids
 from temba.tests.engine import MockSessionWriter
 from temba.tests.s3 import MockS3Client, jsonlgz_encode
 from temba.utils import s3
-from temba.utils.fields import compose_deserialize_attachments, compose_serialize
+from temba.utils.compose import compose_deserialize_attachments, compose_serialize
 from temba.utils.views import TEMBA_MENU_SELECTION
 
 from .tasks import fail_old_messages, squash_msg_counts
